# docs/swagger.yaml
definitions:
  models.Song:
    properties:
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
paths:
  /songs:
    get:
      description: Get songs with optional filtering and pagination.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "500":
          description: Server error
          schema:
            type: string
      summary: Get list of songs
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Add a new song to the library and enrich it with data from an external
        API.
      parameters:
      - description: Song to add
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Add a new song
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Delete a song from the library by its ID.
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Song deleted
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Song not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete a song by ID
      tags:
      - songs
    get:
      description: Get a song from the library by its ID, including the song text.
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Song not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get a song by ID
      tags:
      - songs
    put:
      consumes:
      - application/json
      description: Update the details of a song in the library by its ID.
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      - description: Song data to update
        in: body
        name: song
        required: true
        schema:
          $ref: '#/definitions/models.Song'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Song not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update a song by ID
      tags:
      - songs
swagger: "2.0"